feat(ui): migrate static providers to Schema-Driven UI and stabilize file URIs
