+++ pkg/wallpaper/providers/metmuseum/metmuseum.go
@@ -13,9 +13,6 @@ import (
 	"sync"
 	"time"
 
-	"fyne.io/fyne/v2"
-	"fyne.io/fyne/v2/container"
-	"fyne.io/fyne/v2/widget"
 	"github.com/dixieflatline76/Spice/v2/pkg/i18n"
 	"github.com/dixieflatline76/Spice/v2/pkg/provider"
 	"github.com/dixieflatline76/Spice/v2/pkg/ui/setting"
@@ -98,10 +95,6 @@ func (p *Provider) SupportsUserQueries() bool {
 
 //go:embed MetMuseum.png
 var iconData []byte
-
-func (p *Provider) GetProviderIcon() fyne.Resource {
-	return fyne.NewStaticResource("MetMuseum.png", iconData)
-}
 
 // GetAPIPacing implements the PacedProvider interface to space out API calls.
 func (p *Provider) GetAPIPacing() time.Duration {
@@ -215,7 +208,7 @@ func (p *Provider) FetchImages(ctx context.Context, query string, page int) ([]p
 		}
 
 		// Throttle between batches to respect rate limits (80 req/s, but be gentle)
-		// We use Limit(5) above, so we might hit ~20-50 req/s.
+		// We use Limit(5) above, so we don't burst too hard.
 		// A set sleep ensures we don't burst too hard.
 		time.Sleep(200 * time.Millisecond)
 	}
@@ -512,103 +505,9 @@ func (p *Provider) EnrichImage(ctx context.Context, img provider.Image) (provide
 	return img, nil
 }
 
-// UI Implementation
+// --- UI Implementation (Pure Go) ---
 
-func (p *Provider) CreateSettingsPanel(sm setting.SettingsManager) fyne.CanvasObject {
-	return nil
-}
-
-func (p *Provider) CreateQueryPanel(sm setting.SettingsManager, pendingUrl string) fyne.CanvasObject {
-	// This provider uses the "Museum Template"
-
-	header := wallpaper.CreateMuseumHeader(
-		"The Metropolitan Museum of Art",
-		"New York City, USA",
-		i18n.T("Open Access (CC0)"),
-		"https://www.metmuseum.org/about-the-met/policies-and-documents/open-access",
-		i18n.T("The crown jewel of New York City. From ancient Egyptian temples to modern masterpieces, The Met houses 5,000 years of humanity's greatest creative achievements."),
-		"https://www.google.com/maps/search/?api=1&query=The+Metropolitan+Museum+of+Art",
-		"https://www.metmuseum.org",
-		"https://www.metmuseum.org/donate",
-		sm,
-	)
-
-	// Fixed List of Collections
-	collections := []struct {
-		Name string
-		Key  string
-	}{
-		{i18n.T("Director's Cut: Essential Masterpieces"), CollectionSpiceMelange},
-		{i18n.T("American Wing"), CollectionAmerican},
-		{i18n.T("European Paintings"), CollectionEuropean},
-		{i18n.T("Arts of Asia"), CollectionAsian},
-		{i18n.T("Egyptian Art"), CollectionEgyptian},
-	}
-
-	// Helper to find existing query state
-	getDetails := func(key string) (bool, string) {
-		for _, q := range p.cfg.GetMetMuseumQueries() {
-			if q.URL == key {
-				return q.Active, q.ID
-			}
-		}
-		return false, "" // Not added yet
-	}
-
-	// Create Checkboxes
-	var checks []fyne.CanvasObject
-	for _, col := range collections {
-		col := col // capture
-		active, _ := getDetails(col.Key)
-		dirtyKey := fmt.Sprintf("met_%s", col.Key)
-		callbackKey := fmt.Sprintf("met_cb_%s", col.Key)
-
-		sm.SeedBaseline(dirtyKey, active)
-		chk := widget.NewCheck(col.Name, func(on bool) {
-			if on != sm.GetBaseline(dirtyKey).(bool) {
-				sm.SetSettingChangedCallback(callbackKey, func() {
-					// Actual Save Logic (Deferred)
-					// Fetch fresh ID from config to ensure we target correctly
-					_, cid := getDetails(col.Key)
-
-					if on {
-						if cid != "" {
-							if err := p.cfg.EnableMetMuseumQuery(cid); err != nil {
-								log.Printf("MET: Failed to enable %s: %v", col.Name, err)
-							}
-						} else {
-							desc := fmt.Sprintf(i18n.T("The Met: %s"), col.Name)
-							if _, err := p.cfg.AddMetMuseumQuery(desc, col.Key, true); err != nil {
-								log.Printf("MET: Failed to add %s: %v", col.Name, err)
-							}
-						}
-					} else {
-						if cid != "" {
-							if err := p.cfg.DisableMetMuseumQuery(cid); err != nil {
-								log.Printf("MET: Failed to disable %s: %v", col.Name, err)
-							}
-						}
-					}
-				})
-				// Enable Apply Button
-				sm.SetRefreshFlag(dirtyKey)
-			} else {
-				// Reverted to original state
-				sm.RemoveSettingChangedCallback(callbackKey)
-				sm.UnsetRefreshFlag(dirtyKey)
-			}
-			sm.GetCheckAndEnableApplyFunc()()
-		})
-		chk.Checked = active
-		checks = append(checks, chk)
-	}
-
-	listContainer := container.NewVBox(checks...)
-
-	return container.NewVBox(
-		header,
-		widget.NewSeparator(),
-		widget.NewLabelWithStyle(i18n.T("Collections"), fyne.TextAlignLeading, fyne.TextStyle{Bold: true}),
-		listContainer,
-	)
+// CreateSettingsSchema returns the empty settings definition for the MetMuseum provider.
+func (p *Provider) CreateSettingsSchema() setting.PanelSchema {
+	return setting.PanelSchema{}
 }
